requirements.txt
# Web framework
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
pydantic>=2.0.0

# LLM & embeddings
openai>=1.30.0
httpx>=0.27.0
certifi

# Observability — Langfuse v4
langfuse>=4.0.0
opentelemetry-sdk>=1.24.0
opentelemetry-api>=1.24.0

# Database
#mysql-connector-python>=8.3.0
pymysql

# Config
python-dotenv>=1.0.0

pip-system-certs
